Derive task group priority
We remove this once upon a time, but now I cannot remember why.
In particular, the meaning seems straightforward, in contrast
to the comment in 3a5df350889f5e3f6a060df8722f7780688b6477, that
claimed it was unintuitive.

Adding it back unless we figure out exactly why this shouldn't be
allowed.