Change TaskStatus Semigroup, remove Monoid
- Change TaskStatus Ord s.t.

    NotStarted < InProgress

  This is flipped w.r.t the previous Ord.

- Add special case for

    Completed <> NotStarted == InProgress (and flipped args)

Together, this gives us the algebra we want for deriving a sensible
TaskGroup status:

- Blocked is an ideal (same as before)
- InProgress <> (Completed | NoStarted) == InProgress (new)
- Completed <> NotStarted == InProgress (new)

Notice this means we lose the Monoid, since we have no mempty. Thus it
is removed. Consequently, we swap mconcat for sconcat when deriving
the group status, and we default to the Completed status.

We also add tests for the algebraic laws.