Fix empty task group status/priority deriving
Empty task groups need to be explicitly handled separately, as
mconcat/sconcat does not do what we want (briefly, the failure is
due to the default not being a monoid identity).

Handling empty cases explicitly does what we want.

We also improve the unit tests by testing the SomeTask HasField
instance, which ensures we are actually testing the deriving.