Revamp Index verification
We refactor the index verification logic. The previous logic involved
a series of "verified steps" e.g.

- We'd have an unsafe potentially invariant-breaking function:

    unsafeDoX :: Index -> Index

- We'd then provide a safe wrapper on top:

   doX :: Refined P Index -> Refined P Index
   doX = ... unsafeDoX ...

This worked, except we had a series of refined "doX" functions that
we had to compose together, which many creating many types and many
indirect code (i.e. passing unsafe callbacks). Not only did this make
the code hard to follow, it also included some difficult-to-detect
unsafe functions e.g. we used a '(TaskId -> SomeTask)' callback with
our 'Refined P TaskId', but this was technically unsafe as the callback
could have ignored the Refined TaskId and created its own, destroying
the refinement.

Instead, we refactor 'Index' to 'Index s' where s is verification
state of the index. We now have:

    doX :: Index s -> Index IndexStateUnverified

    verify ::
      (MonadThrow m) =>
      Index IndexStateVerified ->
      m (Index IndexStateUnverified)

Thus instead of painstakingly verifying a series of unsafe steps,
we accept that our modifying functions are unsafe (tracked in the
type), and require a final verify before saving. This yields two
benefits:

1. This new method is much simpler.

2. This final verification is more robust. Now all of our verification
   is in one place, so it is much easier to verify correctness.
   Moreover, the verification as performed in the single 'fromList'
   function is pretty straightforward. It is much easier to understand
   than having several shady steps spread across multiple
   functions.

We have already fixed one bug (inadvertently) by doing this. Previously,
there was no check for a bad blocking reference id when inserting a
task. We now get this for free, since we reuse the one function that
does everything.

This techincally means there is some redundant verification e.g.
we immediately check if a user gives us a bad task id -- to give a
prompt error message -- then we check again at the end. But this is
a small price to pay for the aforementioned benefits.