Revamp Index verification
We refactor the index verification logic. Previously, we'd build
up refined types in stages when inserting tasks e.g.

1. Get TaskId
2. Refined TaskIdNotMember (Index, TaskId)
3. Refined TaskIdNotMember (Index, SomeTask)

Instead of painstakingly creating our proof and mapping it carefully
-- to avoid invalidating -- we opt to instead batch all of our unsafe
operations together, and then do a final "full" verification at the
end before saving. We do this for two reasons.

1. This new method is much simpler. The refined stuff worked, but
   it involved many types as we had to manually pass around our proof
   object. It also made the code quite hard to read, as much of it
   was indirect (e.g. we had to pass around technically unsafe CPS
   functions in order to maintain the refinfements).

2. This final verification is more robust. Now all of our verification
   is in one place, so it is much easier to verifier correctness.
   Moreover, the verification as performed in the single 'fromList'
   function is pretty straightforward. It is much easier to understand
   than having several verification steps spread across multiple
   functions.

We have already fixed on bug (inadvertently) by doing this. Previously,
there was no check for a bad blocking reference id when inserting a
task. We now get this for free, since we reuse the one function that
does everything.

This techincally means there is some redundant verification e.g.
we immediately check if a user gives us a bad task id -- to give a
prompt error message -- then we check again at the end. But this is
a small price to pay for the aforementioned benefits.