Simplify task status <-> text conversions.
We simplify status <-> text conversions by using an Iso. Technically
this is unlawful since (text -> status -> text) is not an isomorphism,
as we add an error message for parsing failures.

But this is fine since we are doing any algebraic reasoning, and it
makes the code easier to understand, having the status <-> text
logic in a single place.